Encapsulate context init creation
