+++ packages/server/src/types/mutations.js
@@ -19,6 +19,7 @@ const mutations = new graphql.GraphQLEnumType({
     JOIN_CHANNEL: { value: "JOIN_CHANNEL" },
     BOT_CREATE: { value: "BOT_CREATE" },
     BOT_UPDATE: { value: "BOT_UPDATE" },
+    GLOBAL_SETTING_UPDATE: { value: "GLOBAL_SETTING_UPDATE" },
   }
 });
